allow @jsonize members to be optional.
The @jsonize attribute may now specify JsonizeOptional.[no,yes] as one of its parameters.
Members default to being JsonizeOptional.no.
Resolves #1 and #2.
There is no strict order to the presence of @jsonize parameters. The following are all valid:
@jsonize int _foo;
@jsonize("foo") int _foo;
@jsonize(JsonizeOptional.yes) int _foo;
@jsonize("foo", JsonizeOptional.yes) int _foo;
@jsonize(JsonizeOptional.yes, "foo") int _foo;
If multiple optional tags are found, the most nested attribute is used:
@jsonize(JsonizeOptional.yes) {
  float bar;
  @jsonize(JsonizeOptional.no) int foo;
}
In the above example, `bar` is optional but `foo` is not.