Add JsonizeOptions with classKey.
Allow users to pass a compile-time configuration to fromJSON in the
form of a JsonizeOptions struct.
Currently it has one field, classKey, which chooses the key used to
perform factory construction.
JsonizeOptions could be extended to configure things like enum
serialization behavior, Nan/Inf handling, and whether to ignore extra
members in objects.